Moved object_manager headers into include/object_manager
