+++ object_manager/include/object_manager/dynamic_object.h
