+++ object_manager/include/object_manager/dynamic_object_xml_parser.h
@@ -7,7 +7,7 @@
 #include <QDebug>
 
 #include <fstream>
-#include "dynamic_object.h"
+#include "object_manager/dynamic_object.h"
 
 
 class DynamicObjectXMLParser {
