+++ object_manager/src/dynamic_object.cpp
@@ -1,4 +1,4 @@
-#include "dynamic_object.h"
+#include "object_manager/dynamic_object.h"
 
 #include <pcl/common/centroid.h>
 #include <pcl/common/transforms.h>
