+++ object_manager/src/load_objects_from_mongo.cpp
@@ -5,7 +5,7 @@
 #include <boost/foreach.hpp>
 #include <pcl_conversions/pcl_conversions.h>
 
-#include "dynamic_object_xml_parser.h"
+#include "object_manager/dynamic_object_xml_parser.h"
 
 #include <QFile>
 #include <QDir>
