+++ object_manager/src/object_manager_main.cpp
@@ -1,4 +1,4 @@
-#include "object_manager.h"
+#include "object_manager/object_manager.h"
 
 int main(int argc, char **argv)
 {
